+++ run.py
@@ -1,0 +1,6 @@
+"""Entry point to sumo application"""
+
+from src.sumo.run import main
+
+if __name__ == '__main__':
+    main()