+++ src/sumo/command_line.py
@@ -155,6 +155,9 @@ def add_run_command_options(subparsers):
     cluster_parser.add_argument('-rep', action='store', type=int, default=RUN_DEFAULTS['rep'], required=False,
                                 help='number of times consensus matrix is created for the purpose of assessing ' +
                                      'clustering quality (default of %(default)s)')
+    cluster_parser.add_argument('-seed', action='store',
+                                  type=int, required=False, default=RUN_DEFAULTS['seed'],
+                                help='random state (none by default)')
 
 
 def add_evaluate_command_options(subparsers):